# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreatedItemDto": {
        "properties": {
          "displayName": { "type": "string" },
          "publicId": { "format": "uuid", "type": "string" },
          "slug": { "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$", "type": "string" },
          "version": { "minimum": 1, "type": "number" }
        },
        "required": ["displayName", "publicId", "slug", "version"],
        "type": "object"
      },
...
      "CreateItemRequestDto": {
        "properties": {
          "attributes": { "additionalProperties": true, "type": "object" },
          "categoryId": { "format": "uuid", "type": "string" },
          "description": { "maxLength": 10000, "nullable": true, "type": "string" },
          "displayName": {
            "example": "Cordless drill",
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "lifecycleStatusId": { "format": "uuid", "type": "string" },
          "storageNodeId": { "format": "uuid", "nullable": true, "type": "string" },
          "tags": {
            "items": { "maxLength": 100, "minLength": 1, "type": "string" },
            "maxItems": 50,
            "type": "array"
          },
          "visibility": {
            "enum": ["authenticated", "private", "public", "unlisted"],
            "type": "string"
          }
        },
        "required": ["categoryId", "displayName", "lifecycleStatusId"],
        "type": "object"
      },
...
      "ItemDetailDto": {
        "properties": {
          "attributes": {
            "additionalProperties": true,
            "description": "Attribute values the actor may read, keyed by stable field key.",
            "type": "object"
          },
          "categoryId": { "format": "uuid", "type": "string" },
          "description": { "nullable": true, "type": "string" },
          "displayName": { "type": "string" },
          "lifecycleStatusId": { "format": "uuid", "type": "string" },
          "publicId": { "format": "uuid", "type": "string" },
          "slug": { "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$", "type": "string" },
          "storageNodeId": { "format": "uuid", "nullable": true, "type": "string" },
          "tags": { "items": { "type": "string" }, "type": "array" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "version": { "minimum": 1, "type": "number" },
          "visibility": {
            "enum": ["authenticated", "private", "public", "unlisted"],
            "type": "string"
          }
        },
        "required": [
          "attributes",
          "categoryId",
          "displayName",
          "lifecycleStatusId",
          "publicId",
          "slug",
          "tags",
          "updatedAt",
          "version",
          "visibility"
        ],
        "type": "object"
      },
...
      "ProblemDetailsDto": {
        "properties": {
          "code": { "example": "VALIDATION_FAILED", "type": "string" },
          "detail": { "example": "One or more fields are invalid.", "type": "string" },
          "fieldErrors": {
            "items": { "$ref": "#/components/schemas/ProblemFieldErrorDto" },
            "type": "array"
          },
          "requestId": {
            "example": "0198f40c-92f3-7a12-bc9a-653f97786c2b",
            "format": "uuid",
            "type": "string"
          },
          "status": { "example": 400, "maximum": 599, "minimum": 400, "type": "number" },
          "title": { "example": "Request validation failed", "type": "string" },
          "type": {
            "example": "https://inventory-atlas.local/problems/validation-failed",
            "type": "string"
          }
        },
        "required": ["code", "detail", "requestId", "status", "title", "type"],
        "type": "object"
      },
      "ProblemFieldErrorDto": {
        "properties": {
          "field": { "example": "displayName", "type": "string" },
          "messages": {
            "example": ["must contain at least 1 character"],
            "items": { "type": "string" },
            "type": "array"
          }
        },
        "required": ["field", "messages"],
        "type": "object"
      },
...
      "UpdatedItemDto": {
        "properties": {
          "displayName": { "type": "string" },
          "invalidators": {
            "description": "Search invalidation registry events this update fired.",
            "items": {
              "enum": ["AttributeChanged", "ItemCreated", "ItemVisibilityChanged"],
              "type": "string"
            },
            "type": "array"
          },
          "publicId": { "format": "uuid", "type": "string" },
          "slug": { "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$", "type": "string" },
          "version": { "minimum": 1, "type": "number" }
        },
        "required": ["displayName", "invalidators", "publicId", "slug", "version"],
        "type": "object"
      },
...
      "UpdateItemRequestDto": {
        "properties": {
          "attributes": {
            "additionalProperties": true,
            "description": "Submitted attribute values keyed by stable field key. Absent keys are kept.",
            "type": "object"
          },
          "categoryId": { "format": "uuid", "type": "string" },
          "description": { "maxLength": 10000, "nullable": true, "type": "string" },
          "displayName": {
            "example": "Cordless drill",
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "expectedVersion": {
            "description": "Known aggregate version; must equal the current Item version.",
            "example": 3,
            "minimum": 1,
            "type": "number"
          },
          "lifecycleStatusId": { "format": "uuid", "type": "string" },
          "storageNodeId": { "format": "uuid", "nullable": true, "type": "string" },
          "tags": {
            "items": { "maxLength": 100, "minLength": 1, "type": "string" },
            "maxItems": 50,
            "type": "array"
          },
          "visibility": {
            "enum": ["authenticated", "private", "public", "unlisted"],
            "type": "string"
          }
        },
        "required": ["expectedVersion"],
        "type": "object"
      },
...
      "VersionConflictProblemDto": {
        "properties": {
          "code": { "example": "VALIDATION_FAILED", "type": "string" },
          "currentVersion": { "example": 8, "minimum": 1, "type": "number" },
          "detail": { "example": "One or more fields are invalid.", "type": "string" },
          "fieldErrors": {
            "items": { "$ref": "#/components/schemas/ProblemFieldErrorDto" },
            "type": "array"
          },
          "requestId": {
            "example": "0198f40c-92f3-7a12-bc9a-653f97786c2b",
            "format": "uuid",
            "type": "string"
          },
          "safeDiff": {
            "additionalProperties": true,
            "description": "Visibility-safe fields that differ between the stored Item and the rejected submission. Each entry carries the current and submitted value; fields the actor cannot view are never included.",
            "example": {
              "displayName": {
                "current": "Cordless drill (workshop)",
                "submitted": "Cordless drill"
              }
            },
            "type": "object"
          },
          "status": { "example": 400, "maximum": 599, "minimum": 400, "type": "number" },
          "title": { "example": "Request validation failed", "type": "string" },
          "type": {
            "example": "https://inventory-atlas.local/problems/validation-failed",
            "type": "string"
          }
        },
        "required": [
          "code",
          "currentVersion",
          "detail",
          "requestId",
          "safeDiff",
          "status",
          "title",
          "type"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "cookie": { "in": "cookie", "name": "inventory_atlas_session", "type": "apiKey" }
    }
  },
  "info": {
    "contact": {},
    "description": "The REST contract consumed by Inventory Atlas clients.",
    "title": "Inventory Atlas API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/items": {
      "post": {
        "operationId": "createItem",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateItemRequestDto" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CreatedItemDto" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          }
        },
        "security": [{ "cookie": [] }],
        "summary": "",
        "tags": ["items"]
      }
    },
    "/api/v1/items/{publicId}": {
      "get": {
        "operationId": "getItem",
        "parameters": [
          {
            "in": "path",
            "name": "publicId",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ItemDetailDto" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          }
        },
        "security": [{ "cookie": [] }],
        "summary": "",
        "tags": ["items"]
      },
      "patch": {
        "operationId": "updateItem",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "description": "Expected aggregate version. Required unless the body carries `expectedVersion`; both must agree when both are present.",
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "publicId",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateItemRequestDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UpdatedItemDto" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VersionConflictProblemDto" }
              }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          }
        },
        "security": [{ "cookie": [] }],
        "summary": "",
        "tags": ["items"]
      }
    },
...
  },
  "servers": [],
...
}